update to 0.2.2
Multiserver update! Can now be played cross-server, we hope.

Little bit of cleanup and stuff also.

Added `players` command. Possibly temporary, possibly will be limited or changed in various ways. Currently lists all a count of awake and asleep players and a list of the server IDs with players in the game, with a player count for each.