Bugfix for players on a server where bot was kicked crashing MUD
Add Player.prototype.getServer() and call that instead of client.guilds.get().

If getting the server fails, knock the player out (.sleep() them) to prevent the server-lookup crash. When they try to do the next command, they'll fail in cantDo() as they're not awake, and won't be able to `joinmud` from the server where bot was kicked, because we already check for that.